requirements.txt
absl-py==2.3.1
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
astunparse==1.6.3
backcall==0.2.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==2.0.12
coloredlogs==15.0.1
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
executing==2.1.0
fast-histogram==0.11
flatbuffers==1.12
gast==0.6.0
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.11.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.2
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest_asyncio==1.6.0
numpy==1.21.6
oauthlib==3.3.1
omegaconf==2.3.0
onnx==1.14.0
onnxoptimizer==0.2.7
onnxruntime==1.14.0
opencv-python==4.5.5.64
opt_einsum==3.4.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.3.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.18.0
python-dateutil==2.9.0
PyYAML==6.0.2
pyzmq==25.1.2
requests==2.27.1
requests-oauthlib==2.0.0
# rknn-toolkit2==1.5.2+b642f30c
rsa==4.9.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.8
scipy==1.5.4
setuptools==75.3.2
six==1.17.0
stack-data==0.6.2
sympy==1.13.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tf-estimator-nightly==2.8.0.dev2021122109
torch==1.10.1
torchvision==0.11.2
tornado==6.4.1
tqdm==4.64.0
traitlets==5.14.3
typing_extensions==4.13.2
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==3.0.6
wheel==0.45.1
wrapt==1.17.2
zipp==3.20.2
